requirements.txt
django-notifications-hq
django-mysql
Pillow